Support sourceSet customization and rev to 2.0.0
Before the build lifecycle was implemented such that if sourceSets were customized,
kotlinter would have already created tasks based on the original sourceSets. If we
do this instead during `project.afterEvaluate` it can work off of the final definitions.

One consequence is that any customization of kotlinter created tasks must be wrapped in an
`afterEvaluate` block since they don't exist during the first build configuration pass. This
prompts a bump to 2.0 since it will be a breaking check for some builds.