Fix support for .editorconfig
Also update the README to reflect that `kotlinter` extension will take precedence over editorconfig when values are specified.

**Why**
Now that `userData` passed into ktlint core takes precedence over values in the `.editorconfig` we need to be use to set a user data key on empty `disabledRules` in the `kotlinter` extension.