Run kotlinter lint and format in classloader isolation
Use the worker classpath to prevent clashes with the configured kotlin compiler version of a project.
This hopefully allows a smoother upgrade path to kotlin 1.4 and fewer compiler version clashes in the future.

One limition is we're doing one worker per sourceSet whereas before that work could be split between workers.
Returning data from workers to tasks is problematic, so we couldn't easily stitch together into a final report. Instead report writing is delegated to the worker.

Since it's a major change, anticipating making the major upgrade to kotlinter and moving all deprecated code.
Also anticipating an upgrade to ktlint 0.38.0 once finalized.